+++ src/main/java/app/control/EnderecoBean.java
@@ -1,5 +1,6 @@
 package app.control;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -10,48 +11,78 @@ import javax.faces.bean.SessionScoped;
 import javax.faces.context.FacesContext;
 
 import app.dao.EnderecoDao;
-import app.model.Documento;
 import app.model.Endereco;
 import app.model.Pessoa;
 
 @ManagedBean
 @SessionScoped
-public class EnderecoBean {
+public class EnderecoBean implements Serializable {
+
+	private static final long serialVersionUID = 1L;
 	private EnderecoDao dao;
 	private Endereco endereco;
-	private Endereco enderecoAnterior;
 	private List<Endereco> enderecos;
 	private Pessoa pessoa;
+	private Endereco enderecoAnterior;
 	private boolean editado;
-
 
 	@PostConstruct
 	public void init() {
-		if(this.dao == null){
-			this.dao = new EnderecoDao(Endereco.class);
+		if (dao == null) {
+			dao = new EnderecoDao(Endereco.class);
 		}
 		this.endereco = new Endereco();
 		this.enderecos = new ArrayList<Endereco>();
 		this.pessoa = new Pessoa();
-
+		this.editado = false;
+		this.enderecoAnterior = new Endereco();
 	}
 
-	public void salvar(Endereco endereco) {
-		this.dao.save(endereco);
+	/**
+	 *
+	 * @param id
+	 * @return
+	 */
+	public String buscarPorId(Long id) {
+		this.endereco = this.dao.findById(id);
+		if (this.endereco != null) {
+			info("Sucesso");
+		} else {
+			warn("Documetno não encotnrado");
+		}
+		return "listarEndereco?faces-redirect=true";
 	}
 
-	public Endereco buscarPorId(Long id) {
-		return this.dao.findById(id);
+	/**
+	 *
+	 * @param id
+	 * @return
+	 */
+	public String remover(Long id) {
+		if (this.enderecos.size() == 1) {
+			error(pessoa.getTipopessoa() + " precisa possuir pelo menos um endereco. Operação não permitida.");
+		} else {
+			Endereco doc = this.dao.remove(id);
+			if (!doc.equals(null)) {
+				this.enderecos.remove(doc);
+				info(doc.getTipo() + " removido com suecsso.");
+			} else {
+				warn("endereco não encontrado na base de dados, favor verificar novamente na lista.");
+			}
+		}
+
+		return "listarEndereco?faces-redirect=true";
 	}
 
-	public Endereco remover(Long id) {
-		return this.dao.remove(id);
-	}
-
+	/**
+	 *
+	 * @param endereco
+	 * @return
+	 */
 	public String atualizar(Endereco endereco) {
 		try {
 			this.endereco = endereco.clone();
-			this.enderecoAnterior.clone();
+			this.enderecoAnterior = endereco;
 			this.editado = true;
 			return "atualizarEndereco?faces-redirect=true";
 		} catch (Exception e) {
@@ -60,13 +91,17 @@ public class EnderecoBean {
 		}
 	}
 
+	/**
+	 *
+	 * @return
+	 */
 	public String salvarAtualizar() {
 		try {
 			this.dao.update(this.endereco);
-			this.enderecos.remove(enderecoAnterior);
-			this.enderecos.add(endereco);
+			System.out.println("Atual \n " + endereco);
+			System.out.println("Anterior \n " + this.enderecos.remove(enderecoAnterior));
 			this.editado = false;
-			info("Endereco atualizado.");
+			info("Endereco " + this.endereco.getTipo() + " atualizados");
 			this.endereco = new Endereco();
 			this.enderecoAnterior = new Endereco();
 			return "listarEndereco?faces-redirect=true";
@@ -80,13 +115,33 @@ public class EnderecoBean {
 		return this.dao.findAll();
 	}
 
-	public void atualizarEnderecos(Pessoa pessoa) {
+	/**
+	 *
+	 * @param pessoa
+	 * @return
+	 */
+	public String atualizarEnderecos(Pessoa pessoa) {
 		this.pessoa = pessoa;
 		this.enderecos = pessoa.getEnderecos();
-		if(this.enderecos.isEmpty() || this.enderecos == null){
-			warn("Nenhum endereço encontrado para " + pessoa.getNome());
+
+		if (this.enderecos.isEmpty() || this.enderecos == null) {
+			warn("Nenhum endereco encontrado para " + pessoa.getNome());
 		}
+		return "listarEndereco?faces-redirect=true";
 	}
+
+	/**
+	 *
+	 * @return
+	 */
+
+	public String cancelarAtualizar() {
+		this.endereco.restaurar(this.enderecoAnterior);
+		this.enderecoAnterior = new Endereco();
+		setEditado(false);
+		return "listarEndereco?faces-redirect=true";
+	}
+
 	public EnderecoDao getDao() {
 		return dao;
 	}
@@ -117,6 +172,14 @@ public class EnderecoBean {
 
 	public void setPessoa(Pessoa pessoa) {
 		this.pessoa = pessoa;
+	}
+
+	public boolean isEditado() {
+		return editado;
+	}
+
+	public void setEditado(boolean editado) {
+		this.editado = editado;
 	}
 
 	/**
